requirements.txt
Flask==3.1.0
Werkzeug>=3.1
waitress>=3.0
gunicorn>=22.0
locust>=2.37
SQLAlchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
Pillow>=10.0
boto3>=1.35
redis>=5.0
rq>=1.16
